fix(JAQPOT-357): fix qsar models asking for s3
qsartoolbox models were asking for the raw model from S3 but this does not make sense since there are no raw models for qsartoolbox, we directly access the API, so this fix will not request for raw models from s3 anymore for qsartoolbox models
