+++ src/main/kotlin/org/jaqpot/api/service/model/QSARToolboxPredictionService.kt
@@ -1,8 +1,7 @@
 package org.jaqpot.api.service.model
 
-import org.jaqpot.api.dto.prediction.PredictionModelDto
+import org.jaqpot.api.entity.ModelType
 import org.jaqpot.api.model.DatasetDto
-import org.jaqpot.api.model.ModelTypeDto
 import org.jaqpot.api.service.qsartoolbox.QSARToolboxAPI
 import org.springframework.stereotype.Service
 
@@ -65,20 +64,19 @@ class QSARToolboxPredictionService(private val qsarToolboxAPI: QSARToolboxAPI) {
     }
 
     fun makePredictionRequest(
-        predictionModelDto: PredictionModelDto,
         datasetDto: DatasetDto,
-        type: ModelTypeDto
+        type: ModelType
     ): List<Any> {
         return when (type) {
-            ModelTypeDto.QSAR_TOOLBOX_CALCULATOR -> {
+            ModelType.QSAR_TOOLBOX_CALCULATOR -> {
                 makeCalculatorPredictionRequest(datasetDto)
             }
 
-            ModelTypeDto.QSAR_TOOLBOX_QSAR_MODEL -> {
+            ModelType.QSAR_TOOLBOX_QSAR_MODEL -> {
                 makeQsarModelPredictionRequest(datasetDto)
             }
 
-            ModelTypeDto.QSAR_TOOLBOX_PROFILER -> {
+            ModelType.QSAR_TOOLBOX_PROFILER -> {
                 makeProfilerPredictionRequest(datasetDto)
             }
