+++ src/main/kotlin/org/jaqpot/api/service/prediction/PredictionService.kt
@@ -1,18 +1,20 @@
 package org.jaqpot.api.service.prediction
 
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import io.github.oshai.kotlinlogging.KotlinLogging
-import org.jaqpot.api.dto.prediction.PredictionModelDto
 import org.jaqpot.api.entity.Dataset
 import org.jaqpot.api.entity.DatasetStatus
+import org.jaqpot.api.entity.Model
 import org.jaqpot.api.mapper.toDto
-import org.jaqpot.api.model.DatasetDto
+import org.jaqpot.api.mapper.toPredictionDto
+import org.jaqpot.api.mapper.toPredictionModelDto
 import org.jaqpot.api.repository.DatasetRepository
 import org.jaqpot.api.service.model.QSARToolboxPredictionService
 import org.jaqpot.api.service.model.dto.PredictionResponseDto
 import org.jaqpot.api.service.model.isQsarModel
 import org.jaqpot.api.service.prediction.runtime.PredictionChain
-import org.springframework.http.HttpHeaders
-import org.springframework.http.MediaType
+import org.jaqpot.api.storage.StorageService
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Service
 import java.time.OffsetDateTime
@@ -22,6 +24,7 @@ import java.time.OffsetDateTime
 class PredictionService(
     private val datasetRepository: DatasetRepository,
     private val predictionChain: PredictionChain,
+    private val storageService: StorageService,
     private val qsarToolboxPredictionService: QSARToolboxPredictionService
 ) {
 
@@ -30,19 +33,12 @@ class PredictionService(
     }
 
     @Async
-    fun executePredictionAndSaveResults(predictionModelDto: PredictionModelDto, dataset: Dataset) {
-        val datasetDto = dataset.toDto()
-
-
-        val headers = HttpHeaders()
-        headers.contentType = MediaType.APPLICATION_JSON
-        headers.accept = listOf(MediaType.APPLICATION_JSON)
-
+    fun executePredictionAndSaveResults(model: Model, dataset: Dataset) {
 
         updateDatasetToExecuting(dataset)
 
         try {
-            val results: List<Any> = makePredictionRequest(predictionModelDto, datasetDto)
+            val results: List<Any> = makePredictionRequest(model, dataset)
             storeDatasetSuccess(dataset, results)
         } catch (e: Exception) {
             logger.error(e) { "Prediction for dataset with id ${dataset.id} failed" }
@@ -71,16 +67,25 @@ class PredictionService(
     }
 
     private fun makePredictionRequest(
-        predictionModelDto: PredictionModelDto,
-        datasetDto: DatasetDto
+        model: Model,
+        dataset: Dataset
     ): List<Any> {
-        if (predictionModelDto.type.isQsarModel()) {
+        val datasetDto = dataset.toDto()
+        if (model.isQsarModel()) {
             return qsarToolboxPredictionService.makePredictionRequest(
-                predictionModelDto,
                 datasetDto,
-                predictionModelDto.type
+                model.type
             )
         }
+
+        val rawModel = storageService.readRawModel(model)
+        val doaDtos = model.doas.map {
+            val rawDoaData = storageService.readRawDoa(it)
+            val type = object : TypeToken<Map<String, Any>>() {}.type
+            val doaData: Map<String, Any> = Gson().fromJson(rawDoaData.decodeToString(), type)
+            it.toPredictionDto(doaData)
+        }
+        val predictionModelDto = model.toPredictionModelDto(rawModel, doaDtos)
 
         val response: PredictionResponseDto =
             predictionChain.getPredictionResults(predictionModelDto, datasetDto)
